- 局内经济系统 - 刷新卡牌、刷新英雄、删卡消耗 - 击杀敌人获得金币

+++ Assets/Scripts/Views/CardView.cs
@@ -292,6 +292,13 @@ public class CardView : MonoBehaviour
     {
         if(!DeleteCardPanelView.Instance.isDeleteCardMode)return;
 
+        if(RuntimeEffectData.coin<Model.deleteCardCost){
+            TipsController.Instance.ShowTips("Not enough coin");
+            return;
+        }
+
+        InGameEconomySystem.Instance.SpendCoin(transform.position,Model.deleteCardCost);
+
         CardSystem.Instance.DeleteCardInBattleField(this);
 
         DeleteCardPanelView.Instance.isDeleteCardMode = false;
